[device-minion-runtime] Improve broadcast message handling - track last received broadcast message number per-minion instead of invalid/race-prone single shared number Normalize log_print console output to use Hxxxx from 0-2112

+++ device-minion-runtime/src/shared/include/message.h
@@ -21,7 +21,7 @@ typedef enum
     MESSAGE_ID_SET_LOG_LEVEL
 } message_id_t;
 
-typedef uint32_t message_number_t;
+typedef uint8_t message_number_t;
 
 typedef struct
 {
